evals 283/284: split mock cluster tools out of the platform-mcp stub
The kubectl_get_nodes / cordon_node distractor tools are NOT part of relay's
platform-mcp (which only exposes the two Slack read tools). Move them out of the
robusta_platform stub into a separate cluster_tools mock MCP server, so the stub
that mirrors platform-mcp carries only the real Slack tools. Behavior is
unchanged; this just stops misattributing the mock cluster tools to platform-mcp.

Mock-only (no live cluster). Verified via tests (3 iterations, opus-4.6 and
opus-4.8): 283 positive reads the channel to disambiguate the node; 284 negative
does not read channel history (deterministic forbidden_tools). Removing the relay
guidance line drops the positive from reading the channel — i.e. the eval gates
on the relay change.
